feat: integrasi stok otomatis dan label sold out di katalog

+++ admin/hapus_menu.php
@@ -1,34 +1,24 @@
 <?php
-
-// memulai session untuk mengakses data login pengguna
 session_start();
-
-// menghubungkan file ke database
 include '../config/koneksi.php';
 
-// proteksi
 if ($_SESSION['role'] != "admin") {
-
-    // arahkan paksa kembali ke halaman login
     header("location:../login.php");
-
-    // menghentikan proses
     exit;
 }
 
-// membersihkan ID dari URL
+// ambil identitas menu dan admin
+$id_stand_admin = $_SESSION['id_stand'];
 $id = mysqli_real_escape_string($conn, $_GET['id']);
 
-// menjalankan query untuk menghapus data menu berdasarkan ID
-$query = mysqli_query($conn, "DELETE FROM menu WHERE id_menu='$id'");
+// eksekusi hapus tapi hanya jika id_menu DAN id_stand cocok
+$query = mysqli_query($conn, "DELETE FROM menu WHERE id_menu='$id' AND id_stand='$id_stand_admin'");
 
-// memberikan feedback kepada user
-if ($query) {
-
-    // jika berhasil, munculkan alert dan pindah ke dashboard
-    echo "<script>alert('Menu berhasil dihapus!'); window.location='dashboard.php';</script>";
+// cek apakah ada baris yang terhapus
+if (mysqli_affected_rows($conn) > 0) {
+    echo "<script>alert('menu berhasil dihapus!'); window.location='dashboard.php';</script>";
 } else {
-
-    // jika gagal, munculkan alert error dan kembali ke dashboard
-    echo "<script>alert('Gagal hapus menu!'); window.location='dashboard.php';</script>";
+    // kalau id nggak cocok atau bukan miliknya
+    echo "<script>alert('gagal hapus!); window.location='dashboard.php';</script>";
 }
+?>
